[Feat][Implement useBook custom hook]
